remove redundant code iin training
